+++ marinechoice/lib/auth/auth_service.dart
@@ -1,15 +1,20 @@
 import 'dart:developer';
 
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class AuthService {
   final _auth = FirebaseAuth.instance;
 
+
   Future<User?> createUserWithEmailAndPassword(
-      String email, String password) async {
+      String email, String password, String name) async {
     try {
       final cred = await _auth.createUserWithEmailAndPassword(
-          email: email, password: password);
+          email: email, password : password);
+      await FirebaseAuth.instance.currentUser!.updateDisplayName(name);
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+      await prefs.setString('username', name);
       return cred.user;
     } catch (e) {
       log(e.toString());
@@ -22,6 +27,9 @@ class AuthService {
     try {
       final cred = await _auth.signInWithEmailAndPassword(
           email: email, password: password);
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+      await prefs.setString('username', cred.user!.displayName!);
+
       return cred.user;
     } catch (e) {
       log("Something went wrong");
@@ -32,6 +40,8 @@ class AuthService {
   Future<void> signOut() async {
     try {
       await _auth.signOut();
+      final SharedPreferences prefs = await SharedPreferences.getInstance();
+      await prefs.remove('username');
     } catch (e) {
       log("Something went wrong");
     }