+++ marinechoice/lib/pages/registerpage.dart
@@ -1,10 +1,13 @@
 import 'dart:developer';
 
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:marinechoice/auth/auth_service.dart';
 import 'package:marinechoice/pages/loginpage.dart';
 import 'package:country_list_pick/country_list_pick.dart';
+import 'package:marinechoice/pages/postpage.dart';
+import 'package:marinechoice/models/user_model.dart';
 
 
 class RegisterPage extends StatefulWidget {
@@ -16,7 +19,6 @@ class RegisterPage extends StatefulWidget {
 class _RegisterPageState extends State<RegisterPage> {
   final _auth = AuthService();
 
-  TextEditingController nameController = TextEditingController();
   TextEditingController usernameController = TextEditingController();
   TextEditingController emailController = TextEditingController();
   TextEditingController passwordController = TextEditingController();
@@ -25,7 +27,6 @@ class _RegisterPageState extends State<RegisterPage> {
   @override
   void dispose() {
     super.dispose();
-    nameController.dispose();
     usernameController.dispose();
     emailController.dispose();
     passwordController.dispose();
@@ -73,17 +74,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   fontWeight: FontWeight.bold,
                 ),
               ),
-              const SizedBox(height: 2),
-              Container(
-                margin: const EdgeInsets.symmetric(horizontal: 50),
-                child: TextField(
-                  controller: nameController,
-                  decoration: const InputDecoration(
-                    hintText: 'Name',
-                  ),
-                ),
-              ),
-              const SizedBox(height: 2),
+              const SizedBox(height: 10),
               Container(
                 margin: const EdgeInsets.symmetric(horizontal: 50),
                 child: TextField(
@@ -93,47 +84,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
                 ),
               ),
-              const SizedBox(height: 2),
-              Container(
-                alignment: Alignment.centerLeft,
-                margin: const EdgeInsets.symmetric(horizontal: 50),
-                child: DropdownButton<String>(
-                  value: _ageinit,
-                  icon: const Icon(Icons.arrow_downward),
-                  onChanged: (String? newValue) {
-                    setState(() {
-                      _ageinit = newValue!;
-                    }
-                    );
-                  },
-                  items: _items().map<DropdownMenuItem<String>>((String value) {
-                    return DropdownMenuItem<String>(
-                      value: value,
-                      child: Text(value),
-                    );
-                  }).toList(),
-                )
-              ),
-              const SizedBox(height: 2),
-              Container(
-                alignment: Alignment.centerLeft,
-                margin: const EdgeInsets.symmetric(horizontal: 30),
-                child: CountryListPick(
-                  theme: CountryTheme(
-                    isShowFlag: true,
-                    isShowTitle: true,
-                    isShowCode: false,
-                    isDownIcon: true,
-                    showEnglishName: true,
-                  ),
-                  onChanged: (CountryCode? code) {
-                    if (kDebugMode) {
-                      print(code!.name);
-                    }
-                  },
-                ),
-              ),
-              const SizedBox(height: 2),
+              const SizedBox(height: 10),
               Container(
                 margin: const EdgeInsets.symmetric(horizontal: 50),
                 child: TextField(
@@ -143,7 +94,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
                 ),
               ),
-              const SizedBox(height: 2),
+              const SizedBox(height: 10),
               Container(
                 margin: const EdgeInsets.symmetric(horizontal: 50),
                 child: TextField(
@@ -154,14 +105,14 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
                 ),
               ),
-              const SizedBox(height: 2),
+              const SizedBox(height: 10),
               Container(
                 margin: const EdgeInsets.symmetric(horizontal: 50),
                 child: TextField(
                   controller: confirmPasswordController,
                   obscureText: true,
                   decoration: const InputDecoration(
-                    hintText: 'confirm password',
+                    hintText: 'Confirm password',
                   ),
                 ),
               ),
@@ -222,10 +173,11 @@ class _RegisterPageState extends State<RegisterPage> {
   }
 
   _signup() async{
-    final user = await _auth.createUserWithEmailAndPassword(emailController.text, passwordController.text);
+    final user = await _auth.createUserWithEmailAndPassword(emailController.text, passwordController.text, usernameController.text);
 
     if(user != null) {
       log("User Created Successfully");
     }
   }
+
 }